QtOutputFormatter: Allow standalone file:// links
Either alone on a line, or with text before, or with
text afterwards, separated by colon, space, tab, or
brace as before.

Especially thought to capture cases like

 file:///home/user/main.cpp:157

on a single line, but accept that also as long as
the file link is separated by a word separator from
the sourrounding text.

Fixes: QTCREATORBUG-20406
Change-Id: I25bceb186818c9ea680741573a806f66395c3e16
Reviewed-by: Samuel Gaist <samuel.gaist@idiap.ch>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>