Recognize UNC paths as absolute (i.e. not relative)
IoUtils::isRelativePath() didn't attempt to consider UNC paths, due to
a belief that qmake fails on them so badly that it wasn't worth the
extra code.  However, it turns out Qt Creator's copy of this code does
need to take this into account, so start the change off in qmake's
version so as to keep in sync.

This is a cherry pick from qtbase commit 4b4a288f5f88077d647d47e9bdd911edf6cdca72

Fixes: QTCREATORBUG-21881
Written-by: Edward Welbourne <edward.welbourne@qt.io>
Change-Id: I3084b87c1d3ca6508255e94e04ac8db3ceaebb7e
Reviewed-by: Nikolai Kosjar <nikolai.kosjar@qt.io>