Fix cloning DesktopQmakeRunConfiguration
The value of a terminal aspect is not exported in toMap() unless it was
set by the user, so we need help from the project manager to get the
default value. The QbsRunConfiguration already works this way.
Maybe the terminal aspect could export all of its information and we
could drop the fromMap() override from both run configs, but that should
be a different patch.

Fixes: QTCREATORBUG-21918
Change-Id: I712158f4bcb4dbd3827568dd4e1fe8b21d802f47
Reviewed-by: Orgad Shaneh <orgads@gmail.com>