ProjectExplorer: Sanitize the "Add run config" UI
This functionality was implemented via a pop-up menu, which was close to
unusable if more than a few candidates existed (for example: Qt Creator
with autotests enabled).
We now use a proper dialog. The list of candidates is sortable, can be
filtered and includes information about which project file the target
executable comes from.

Fixes: QTCREATORBUG-19955
Change-Id: Ife087ad69a7e43e280d13c528d21f94a1ae48d4d
Reviewed-by: hjk <hjk@qt.io>