ProjectExplorer: Avoid querying network compilers
...in case the network compiler is the first in PATH.

If e.g. /usr/lib/icecc/bin comes first in system PATH and it's the
compiler being auto-detected findLocalCompiler() returned
/usr/lib/icecc/bin instead of the real local compiler.

The "compilerPath" that findLocalCompiler() got was always additionally
prepended to the "env" it got. The additionally prepended one was
successfully skipped, but then it run into the original one.

Fix this by ensuring that findLocalCompiler() filters out all network
compilers and not only the first one.

Change-Id: Ia44eef67497d760714d72e8a28620f423c2860d2
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>