Allow pinning files to ensure that they are always open
This patch allows pinning files within a session. Pinning a file puts
it at the top of the Open Documents list, and prevents Close All
from closing it until it is unpinned.

This is useful for files that should always be open for a given
session.

[ChangeLog] Files can now be pinned via the context menu.
Pinning a file keeps it at the top of the Open Documents list,
and prevents Close All and similar actions from closing it
until it is unpinned. This provides a way to quickly close any
open files without closing important ones.

Change-Id: If47a599fb272db4c78a71eabe6fb29215a9a8a11
Fixes: QTCREATORBUG-21899
Reviewed-by: Eike Ziller <eike.ziller@qt.io>