Properly import qmake projects
The code simply picked one matching toolchain and set it on the
temporary kit, ignoring the fact that we have pairs of toolchains (C and
C++). So the second one was set to some random value and typically did
not match the first one, resulting in a broken Kit.
We now collect both types of toolchains.

Fixes: QTCREATORBUG-18153
Fixes: QTCREATORBUG-19423
Change-Id: I348e5467f30d449cc344946ec7e87d4ca5ddfbd1
Reviewed-by: Christian Stenger <christian.stenger@qt.io>