Output pane manager: Fix two inconsistencies
1) Currently deselected panes that popped up due to some event behaved
differently with respect to their button visibility: For instance, if
the user hides the application output pane and then new application
output arrives (and the respective item in the "Build & Run" settings is
enabled), then the pane widget would appear as well as the pane button.
In contrast, for the issues pane only the widget itself would appear,
but the button was still hidden.
Now the button appears along with the widget for all output panes.
2) When the deselected app output pane popped up due to new output (as
described above), then the check button in the output pane menu failed
to update, i.e. was still unchecked and thus inconsistent with the
actual state. Now it is in sync.

Change-Id: I5e100363c6c01327b2c03b7f22ca173d41d13d7c
Reviewed-by: Eike Ziller <eike.ziller@qt.io>