Remove ProjectExplorerPlugin::setBuildDirectoryTemplate()
The default build directory template is part of the
ProjectExplorerSettings data structure, but was read and written
separately by the settings page, which was confusing, because it meant
the respective value was essentially written twice, once with a dummy
value and then with the real one right afterwards. Fix that.

Change-Id: Id9536560dbd0505c2099abe627bca7e116fec296
Reviewed-by: hjk <hjk@qt.io>