Rewrite qdump__std____1__map(d, value)
Rewrote debugger helper `qdump__std____1__map(d, value)` which was previously not working on most platforms that I've seen (I have not seen any platforms work, but have not tested all. Was crashing previously trying to get members that don't exist from the tree resulting in <not accessible> message on debugger.

This rewrite has been proven to work on Macs so far.