Help: Improve handling of results from index lookup
If help is only found by looking up in the index, show a selection
dialog to the user even if there is only one result. Otherwise we create
the impression that we really think that the help we find is the correct
one.

Also do not add the help text to the tool tip in this case.

Test case: struct Foo { static void objectCreated() {} };

Change-Id: I9579302843ea2923e06f56f4b646dd101f183b3f
Reviewed-by: David Schulz <david.schulz@qt.io>