Fix soft assert when opening widget designer form
Creating a designer form adds it to design mode's editor tool bar, which
is correct. But this should not try to switch the drop down to this new
document, which has only just been created and is not registered
anywhere. It is not even officially the current document yet. Switching
the drop down is done when it becomes the current document anyhow.

Change-Id: If7fa8273600d83eda2cd1fa4731924007f8ca1f5
Fixes: QTCREATORBUG-20270
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>