Fix collapsed state when opening file while Projects mode is open
The editor manager switches to edit mode when opening a file while a
mode without an editor area is open. Project explorer saves settings
when the mode is switched away from Projects mode, but did that before
the editor's state was restored by the editor manager.

Settings saving should not be done while transitioning states, so delay
that to a clean, next event loop processing.

Fixes: QTCREATORBUG-16982
Change-Id: I165ed2f84611f45341f81d24a12dc257e208ef1f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>