Windows: Fix missing debug info for qtcreator executable
When we set a VERSION, there is QTBUG-74265 which leads to wrong rules
for the pdb file.
On Windows we do not really need the VERSION for the qtcreator
executable (we create an .rc file for the executable details for
Windows), so unset the VERSION that is set in qtcreator.pri on Windows.

Change-Id: I1e0289fe027b48560ec71f6df70522174770fe84
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>