Windows: Fix submenu arrow style
Since dccf5a5ae7cda289be1796f145742ccfe478f086 the submenus have the
parent menu as an actual parent widget. That makes ManhattanStyle handle
them differently, because they are now children of the menu bar, and
therefore considered "panel widgets" to be styled.

An undesired side-effect is that ManhattanStyle decides to style the
submenu arrows itself then.

The patch makes ManhattanStyle not style arrows if they are painted as
part of a menu.

Fixes: QTCREATORBUG-21376
Change-Id: Ia5556c7327c3f0c2f21cadc86047bec785da27df
Reviewed-by: Robert Loehning <robert.loehning@qt.io>