bare-metal: Search the include paths of IAR toolchain
Seems, the IAR compiler has not an explicitly options to show
a list of system include paths. But, we use the following trick
to enumerate this directories. We specify the '--preinclude'
option with the wrong value (e.g. a dot). In this case the
compiler fails and its error output contains a mention about
the using search paths in a form of tokens:

 ' searched: "<path/to/include>" '.

Where are the resulting paths are escaped with a quotes.

Change-Id: Ie9cbf08ae84971bfc8ea4b1cd2ffe3385324710f
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>