Add more Python features
On the wizard related to Main Windows, not there are more
options to specify the name of the files and classes.

It's possible to Add and Remove files on Python projects
that use the new .pyproject file.

Improved the empty application,
to template the .pyproject file and use the proper name
for the file.

Added the icons for file overlay.

Change-Id: Iaba7feda69e0f608260b5fb1d1b04b2a42b08c2d
Reviewed-by: hjk <hjk@qt.io>