Advanced Search: Add horizontal scroll bar
Compute the actual size hint for the search results and make the column
resize to its contents.
To make the highlight of selected items still span the whole widget, we
force the column to span at least the width of the tree view, by setting
the minimum section size accordingly on resize events.

Fixes: QTCREATORBUG-21099
Change-Id: Ia9e88fba8185429fede8a82402a62285afb1e0dd
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: David Schulz <david.schulz@qt.io>