QmlDesigner: Workaround for regression in Qt
Editing finished is emitted, when we show an error dialog,
and the TextFiled looses focus. Even if the value did not change.

This triggers a second error dialog and hangs on macOS.
This is a regression from Qt 5.11 to Qt 5.12.x.

The workaround is to check for the value change manually.

Task-number: QDS-519
Change-Id: Ie8a3f342bc0b81941d07773d52947ad76ae27205
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>