QmlDesigner: Do not flag files as changed when just opening
If a file without size is opened we annotate a default size.
Technically this is changing the file because when saving the file
we will generate an annoation like this:

/*##^## Designer {
    D{i:0;autoSize:true;height:480;width:640}
}
 ##^##*/

But since this annoation is done automatically without user interaction,
we should not mark the file as dirty in this case.

Task-number: QTCREATORBUG-22216
Change-Id: I32686baf976927c3edaf12b07ba81c62759633c0
Reviewed-by: Robert Loehning <robert.loehning@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>