QmlDesigner: Use AdaptiveDecimalStepType in timeline DoubleSpinBox
This is something we might consider for the Qt Quick Controls,
since we have controls where a pre defined step size does not
make much sense, or we simply do not know the step size.

Adaptive decimal step means that the step size will continuously
be adjusted to one power of ten below the current value.
So when the value is 1100, the step is set to 100,
so stepping up once increases it to 1200. For 1200
stepping up takes it to 1300. For negative values,
stepping down from -1100 goes to -1200.

Task-number: QDS-534
Change-Id: I18d3cb43e4fa140d94336b5128c4f1f6707880dd
Reviewed-by: Brook Cronin <brook.cronin@qt.io>
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>