Scroll Qt Quick Application Wizard: Enable minimal Qt version selection
This change makes the combobox for the minimal Qt Version in the
"Scroll" Qt Quick Application Wizard visible.

In contrast to the other Qt Quick Application Wizards, the "Scroll"
variant had the minimal Qt version selection hidden. That was OK in the
initial wizard commit, where only one Qt Version would have been
selectable. But with the subsequent additions of Qt Versions in the
wizard.json, the combobox visibility should have been changed to true.

The consequence of the invisible combobox was that the default was
secretly selected. The default was Qt 5.10 for a while. But the default
changed recently with 52f95bae8b6a7c38b2d183522e1a7e550509d52f and
caused QTCREATORBUG-22245.

Task-number: QTCREATORBUG-22245
Change-Id: If4d2f384a28bf3dd130b097022ea6d8d30ef7509
Reviewed-by: Eike Ziller <eike.ziller@qt.io>