Qnx: Use SimpleRunWorkerFactory instead of manual registration
Looks like the extra constraint was effectively checking only for the
presence of a QnxRunConfiguration, which are only created for devices of
type Constants::QNX_QNX_OS_TYPE, which happen to be the only one with
with a bool(...dynamicCast<const QnxDevice>()) == true.

The previous runConfig->isEnabled() is unsystematic, not used anywhere
else, and effectively globally handled in PE::canRunStartupProject.

Change-Id: I5bf814fb5316bb38d315179b16ad4a78faff6899
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>