CPaster: Remove KDE paster
There is no official API anymore after the latest change of
the KDE paster and it was probably almost no more in use
after the need for credentials.
Remove everything related except the base implementation
of the sticky notes part as it could be re-used for another
paster that based on the old approach used by the KDE paster.

Change-Id: I0e2c3f279bc9eaa373147ee2909c9538f0d62498
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>