ProjectExplorer: Remember build step expansion state
... when switching targets.
If a user expands a build step and then switches the target, it's
reasonable for them to assume that the step is still expanded when
switching back.
Ideally, we would also do this for aspects (if only for consistency),
but unlike for build steps there's no central code location, so let's
leave it at this for now.

Fixes: QTCREATORBUG-21343
Change-Id: Ic142ab5d676bf7d3b6aeabb6f7a26be7bd18f22c
Reviewed-by: hjk <hjk@qt.io>