QmakeProjectManager: Remove python file type
This is a partial regression from b503ea7f06.

The main goal was just to have an overlay on the Python files,
but since qmake and python don't really interact, adding a new
FileType triggered all the contents from "Other sources" to
be duplicated in the "Python sources" node.

The Python sources node is not necessary at the moment,
so I remove it to avoid duplicated tree elements.

Change-Id: I61e9e51327689636328a746059159c962270e805
Fixes: QTCREATORBUG-22292
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>