Fix passing of kit IDs for subproject to JSON wizards
The list of preselected kits in case of sub projects
was passed in a format not compatible with wizard variables
(QList<Id>, but only QString and QStringList are supported).

This fixes the visibility of the Kit selection page of the "Empty qmake
Project" and "Auto Test Project" wizards when triggered from "Add New
Subproject" from the project tree's context menu.

Change-Id: Ica7305825d6323697c9b0788a9634f3d806b9d50
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>