QmakePM: Only consider the first make step for subproject builds
When the user requests to build a sub-project or a single file of a
qmake project, we simply run the normal steps from the build
configuration, and the make steps modify themselves to run in the
respective subdirectory and/or with specific targets. Normally, there is
only one such make step. If the user added more make steps, it is very
likely that these only make sense for a top-level build, so we simply
make them no-ops when building a sub-project or single file.

Fixes: QTCREATORBUG-15794
Change-Id: I960cf60dba5e5d6b6e839d35e34650427f659110
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>