ProjectExplorer: Fix renaming files in resources
This amends commit e161b5d1cd, which fixed the problem only for top-
level entries in the resource file.
The underlying problem is that we have two different classes
ResourceFolderNode and SimpleResourceFolderNode, which look suspiciously
similar, with most of the latter's base class overrides being exact
copies of the former's. For now, we just get rid of these copies. At
some later point, we might want to drop one of these classes altogether.

Fixes: QTCREATORBUG-22313
Change-Id: I80094ca12f4e534daf40ca2f3cbf01a0ae4808d9
Reviewed-by: hjk <hjk@qt.io>