PerfProfiler: Remove backwards compatibility support
Define a new file extension, new file magic, and remove support for the
old data format. You won't be able to load traces generated with older
Qt Creators in 4.10 anymore and you won't be able to use older
perfparsers with Qt Creator 4.10.

On the flip side, we can move along more quickly now and integrate
features from KDAB's perfparser. This is to be preferred because the
maintenance of two different perfparsers is quite a burden.

Change-Id: I69918eadc3294abfbb0b809a54c3c9f25405409a
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>