LSP: reredo completion replacement logic
replacing the text from word start to current cursor position only works
if we are not currently at a word start. For example completing:

  object.function(var|)

will not replace the already typed var, because selecting from cursor
position to word start does not select anything.

Try selecting from current position backwards as long as the inserted
text matches case insensitive to the already available text, or select
the match of the "[a-zA-Z_][a-zA-Z0-9_]" regex. Whichever selected text
is longer will be replaced.

Change-Id: I73965183ba430a8de1b7725e1f7bc7621861433f
Reviewed-by: Christian Stenger <christian.stenger@qt.io>