Fix error message when adding empty qmake project as subproject
The JSON wizard code expects that at least one file gets opened either
as a project or in an editor. This was not true when adding an empty
qmake project, because its .pro file loses the "open as project"
property when added as a sub-project, and there are no other files to
open.
We fix this by also setting "open in editor" for the .pro file, which
makes sense (also for the top-level case), because a user is likely to
add custom content there right away.

Change-Id: Icdf68bccb2488bdb26b6b3c8410da901a4eb32df
Reviewed-by: Eike Ziller <eike.ziller@qt.io>